+++ CHANGELOG.md
@@ -17,6 +17,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 - `gimme maelstrom` works with or without xgboost (but will give a warning without xgboost) 
 - fixed warning "in validate_matrix(): Row sums in df are not close to 1. Reormalizing rows..."
+- fixed multiprocess.Pool Warnings
+- silence warnings from external tools in motif prediction (`pp_predict_motifs()` to be exact)
 
 ## [0.17.2] - 2022-10-12